edit readme to include line about os compatibility
merge this commit, guys, cause I spent 2-3 hours trying to run/debug cli commands on windows and noticed only after that it is intended for use on mac/linux only. And guess I'm not the only one considering comments on one of opened issues, so until it doesn't work on windows, make that info visible to all devs in first place. Thanks